[Hotfix] int double 자료형 안맞는 버그수정

+++ Android/app/src/main/java/com/example/decisionroulette/api/vote/VoteDto.kt
@@ -15,7 +15,7 @@ data class VoteListItem(
  */
 data class VoteItem(
     val name: String,
-    val voteRate: Int // 백분율 (0 ~ 100)
+    val voteRate: Double // 백분율 (0 ~ 100)
 )
 
 /**
